requirements.txt
numpy~=1.23.2
torch~=1.12.1+cu113
pandas~=1.4.3
matplotlib~=3.5.3
PyYAML~=6.0